requirements.txt
alembic==1.1.0
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
importlib-metadata==0.20
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.4
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
mysqlclient==1.4.4
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.2
redis==3.3.8
six==1.12.0
SQLAlchemy==1.3.8
vine==1.3.0
Werkzeug==0.15.5
WTForms==2.2.1
zipp==0.6.0
